+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@nextui-org/react": "1.0.0-beta.13",
+    "framer-motion": "^10.15.0",
     "localforage": "^1.10.0",
     "match-sorter": "^6.3.1",
     "react": "^18.2.0",
