+++ src/pages/cart/profile.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-
+import { ProfileProps } from '../../components/profile/profileInterface'
 function Profile() {
   return (
-    <div><p>This is cart page.</p>
-    <Link to='/profile'>GO HOME</Link>
+    <div><p>This is Profile page.</p>
+    <Link to='/'>GO HOME</Link>
     </div>
   )
 }
