+++ src/pages/cart/settings.tsx
@@ -3,8 +3,8 @@ import { Link } from 'react-router-dom'
 
 function Settings() {
     return (
-        <div><p>This is cart page.</p>
-            <Link to='/settings'>GO HOME</Link>
+        <div><p>This is settings page.</p>
+            <Link to='/'>GO HOME</Link>
         </div>
     )
 }
